backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
numpy
scikit-learn
